refactor(core): extract dual-mono detection from audio decode
Move mix_channels_to_mono, is_dual_mono, and DualMonoDetector into
decode/dual_mono.rs. Public decode_audio_* paths are unchanged.
